requirements.txt
aiogram==3.2.0
ujson
aiohttp[speedups]
pillow
